Rough Code, many changes
Main.py is finally included.
Functionality was broken on the previous way of working.
Many things need to be cleaned up in the old files, as well as main.py.
Main.py adds the GUI, it's a mess, just wanted to see it operational first before I cleaned up the old way to work with the new way.
I currently use .trace on some StringVars, I don't like it , maybe 'bind', I think I'll change it after some code reorganization.

There needs to be feedback to the user so they know that it's working.

Now the ReadMe needs changes to reflect it's current state.